refactor: RacingGameView에서 InputValidator를 분리해 하나의 클래스로 만듬
