+++ src/main/java/racingcar/view/RacingGameView.java
@@ -11,8 +11,10 @@ import racingcar.domain.Car;
 
 public class RacingGameView {
 	private static final RacingGameView racingGameView = new RacingGameView();
+	private final InputValidator inputValidator;
 
 	private RacingGameView() {
+		inputValidator = InputValidator.getInputValidator();
 	}
 
 	public static RacingGameView getRacingGameView() {
@@ -24,7 +26,7 @@ public class RacingGameView {
 		do {
 			System.out.println(CAR_NAME_INPUT_MESSAGE);
 			carNamesList = new ArrayList<>(Arrays.asList(Console.readLine().split(",")));
-		} while (checkNameExceptions(carNamesList));
+		} while (inputValidator.checkNameExceptions(carNamesList));
 		return carNamesList;
 	}
 
@@ -33,7 +35,7 @@ public class RacingGameView {
 		do {
 			System.out.println(REPETITION_NUMBER_MESSAGE);
 			repetitionNumber = Console.readLine();
-		} while (!checkNotNaturalNumberException(repetitionNumber));
+		} while (!inputValidator.checkNotNaturalNumberException(repetitionNumber));
 		return Integer.parseInt(repetitionNumber);
 	}
 
@@ -50,45 +52,5 @@ public class RacingGameView {
 
 	public void printWinners(List<String> winners) {
 		System.out.println(FINAL_RESULT_MESSAGE + String.join(", ", winners));
-	}
-
-	private boolean checkNameExceptions(List<String> carNamesList) {
-		return !checkLongNameException(carNamesList) || checkBlankNameException(carNamesList);
-	}
-
-	private boolean checkLongNameException(List<String> carNamesList) {
-		try {
-			if (carNamesList.stream().anyMatch(carName -> carName.length() > MAXIMUM_NAME_LENGTH)) {
-				System.out.println(LONG_CAR_NAME_INPUT_ERROR_MESSAGE);
-				throw new IllegalArgumentException();
-			}
-		} catch (Exception e) {
-			return false;
-		}
-		return true;
-	}
-
-	private boolean checkBlankNameException(List<String> carNamesList) {
-		try {
-			if (carNamesList.stream().anyMatch(carName -> carName.length() == 0)) {
-				System.out.println(BLANK_CAR_NAME_INPUT_ERROR_MESSAGE);
-				throw new IllegalArgumentException();
-			}
-		} catch (Exception e) {
-			return false;
-		}
-		return true;
-	}
-
-	private boolean checkNotNaturalNumberException(String repetitionNumber) {
-		try {
-			if (repetitionNumber.chars().anyMatch(number -> !Character.isDigit(number))) {
-				System.out.println(REPETITION_NUMBER_ERROR_MESSAGE);
-				throw new IllegalArgumentException();
-			}
-		} catch (Exception e) {
-			return false;
-		}
-		return true;
 	}
 }
